rag-enterprise/backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-cohere
langchain-pinecone
cohere
pinecone
pypdf
python-multipart
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt<4
argon2-cffi
python-jose[cryptography]
google-auth
requests
email-validator
slowapi
supabase